pyproject.toml
[project]
name = "vidbot"
version = "0.1.0"
description = "A YouTube video chatbot with transcript analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "youtube-transcript-api>=1.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langchain-huggingface>=0.0.6",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "beautifulsoup4 ",
    "langchain-text-splitters"
]



